pocketguard/requirements.txt
fastapi
uvicorn[standard]
PyGithub
python-dotenv
httpx
celery
redis
flake8
